checkpoint

+++ logout.go
@@ -14,6 +14,10 @@ func (nbrew *Notebrew) logout(w http.ResponseWriter, r *http.Request) {
 	if !ok {
 		logger = slog.Default()
 	}
+	if nbrew.DB == nil {
+		http.Error(w, "404 Not Found", http.StatusNotFound)
+		return
+	}
 	authenticationTokenHash := getAuthenticationTokenHash(r)
 	if authenticationTokenHash == nil {
 		http.Redirect(w, r, nbrew.Protocol+nbrew.AdminDomain+"/admin/", http.StatusFound)
@@ -50,17 +54,19 @@ func (nbrew *Notebrew) logout(w http.ResponseWriter, r *http.Request) {
 			Value:  "",
 			MaxAge: -1,
 		})
-		_, err := sq.ExecContext(r.Context(), nbrew.DB, sq.CustomQuery{
-			Dialect: nbrew.Dialect,
-			Format:  "DELETE FROM authentication WHERE authentication_token_hash = {authenticationTokenHash}",
-			Values: []any{
-				sq.BytesParam("authenticationTokenHash", authenticationTokenHash),
-			},
-		})
-		if err != nil {
-			logger.Error(err.Error())
-			http.Error(w, messageInternalServerError, http.StatusInternalServerError)
-			return
+		if authenticationTokenHash != nil {
+			_, err := sq.ExecContext(r.Context(), nbrew.DB, sq.CustomQuery{
+				Dialect: nbrew.Dialect,
+				Format:  "DELETE FROM authentication WHERE authentication_token_hash = {authenticationTokenHash}",
+				Values: []any{
+					sq.BytesParam("authenticationTokenHash", authenticationTokenHash),
+				},
+			})
+			if err != nil {
+				logger.Error(err.Error())
+				http.Error(w, messageInternalServerError, http.StatusInternalServerError)
+				return
+			}
 		}
 		http.Redirect(w, r, nbrew.Protocol+nbrew.AdminDomain+"/admin/", http.StatusFound)
 	default:
